Mempool tests moving to mempool project (#1241)
